requirements.txt
attrs >= 18.1.0
# docker
python-dateutil
redis
pyyaml

# For distributed queues

celery

# For testing

pytest
interruptingcow
